+++ src/DTO/OrderData.php
@@ -99,6 +99,16 @@ class OrderData implements JsonSerializable
      */
     private $deliveryPhoneNumber;
 
+    /**
+     * @var string
+     */
+    private $shippingStreetAndNumber;
+
+    /**
+     * @var string
+     */
+    private $billingStreetAndNumber;
+
     public function __construct(
         Amount $amount,
                $redirectUrl,
@@ -357,6 +367,26 @@ class OrderData implements JsonSerializable
         $this->payment = $payment;
     }
 
+    public function getShippingStreetAndNumber(): string
+    {
+        return $this->shippingStreetAndNumber;
+    }
+
+    public function setShippingStreetAndNumber(string $shippingStreetAndNumber): void
+    {
+        $this->shippingStreetAndNumber = $shippingStreetAndNumber;
+    }
+
+    public function getBillingStreetAndNumber(): string
+    {
+        return $this->billingStreetAndNumber;
+    }
+
+    public function setBillingStreetAndNumber(string $billingStreetAndNumber): void
+    {
+        $this->billingStreetAndNumber = $billingStreetAndNumber;
+    }
+
     public function jsonSerialize()
     {
         $lines = [];
@@ -371,7 +401,7 @@ class OrderData implements JsonSerializable
             ],
             'billingAddress' => [
                 'organizationName' => $this->cleanUpInput($this->getBillingAddress()->company),
-                'streetAndNumber' => $this->cleanUpInput($this->getBillingAddress()->address1),
+                'streetAndNumber' => $this->cleanUpInput($this->getBillingStreetAndNumber()),
                 'city' => $this->cleanUpInput($this->getBillingAddress()->city),
                 'postalCode' => $this->cleanUpInput($this->getBillingAddress()->postcode),
                 'country' => $this->cleanUpInput(Country::getIsoById($this->getBillingAddress()->id_country)),
@@ -381,7 +411,7 @@ class OrderData implements JsonSerializable
             ],
             'shippingAddress' => [
                 'organizationName' => $this->cleanUpInput($this->getShippingAddress()->company),
-                'streetAndNumber' => $this->cleanUpInput($this->getShippingAddress()->address1),
+                'streetAndNumber' => $this->cleanUpInput($this->getShippingStreetAndNumber()),
                 'city' => $this->cleanUpInput($this->getShippingAddress()->city),
                 'postalCode' => $this->cleanUpInput($this->getShippingAddress()->postcode),
                 'country' => $this->cleanUpInput(Country::getIsoById($this->getShippingAddress()->id_country)),