+++ changelog.md
@@ -9,6 +9,7 @@
 + Changed how secure key is generated to avoid errors.
 + Fixed issue where payment method save with multishop would break another shop payment method settings.
 + Fixed translations that were not working because of php.
++ Improved address by adding address2 part to mollie API if it is added to address.
 
 ## Changes in release 5.2.1 ##
 + Fixed issue where orders would have duplicated lines in list if there are more than 1 transaction.